Adding AppState For MAin GAme
Export all of Main functionality except Health to AppState lel
GAme Ends After 10 Enemys Passing Through
