+++ src/mygame/IllegalImmigrant.java
@@ -39,18 +39,18 @@ public class IllegalImmigrant extends AbstractControl {
     public Vector3f sphereDirection;
 
     public IllegalImmigrant(WaveSpawner w) {
-        m = Main.instance.getTdMap();
+        m = MainGame.instance.getTdMap();
         xn = m.getnodex();
         yn = m.getnodey();
         this.speedFactor = 30.0f;
         velocity = new Vector3f(0, 0, 0);
         nrCheckpoints = xn.length;
-        bgObject = Main.instance.getFloor();
+        bgObject = MainGame.instance.getFloor();
         ratioxr = bgObject.getWidth() / m.getWidth();
         ratioyr = bgObject.getHeight() / m.getHeight();
         Vector3f spawn = new Vector3f(-(bgObject.getWidth() / 2) + xn[0] * ratioxr, -(bgObject.getHeight() / 2) + yn[0] * ratioyr, 0.0f);
         geom = GeometryCreator.instance.createSphere(spawn);
-        Main.instance.attachToRootNode(geom);
+        MainGame.instance.attachToRootNode(geom);
         geom.addControl(this);
         this.w = w;
